// Frontend/src/app/core/services/aminety-service/models/BookingResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BookingResponseDto = {
    bookingId?: string;
    userId?: string;
    amenityName?: string | null;
    slotType?: string | null;
    slotLabel?: string | null;
    slotDate?: string;
    startTime?: string;
    endTime?: string;
    peopleCount?: number;
    status?: string | null;
    bookedAt?: string;
    cancelledAt?: string | null;
};