// Frontend/src/app/core/services/aminety-service/models/BookingListResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BookingResponseDto } from './BookingResponseDto';
import type { PaginationDto } from './PaginationDto';
export type BookingListResponseDto = {
    data?: Array<BookingResponseDto> | null;
    pagination?: PaginationDto;
};